Updated Unit tests

+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/pet/PetController.java
@@ -17,7 +17,7 @@ public class PetController {
     }
 
     @GetMapping("/{petId}")
-    public PetDTO getPet(@PathVariable int petId) {
+    public PetDTO getPet(@PathVariable long petId) {
         throw new UnsupportedOperationException();
     }
